CustomerSite added function finished
