+++ src/main/java/cn/stylefeng/guns/modular/company/controller/CustomerSiteController.java
@@ -13,44 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cn.stylefeng.guns.modular.system.controller;
+package cn.stylefeng.guns.modular.company.controller;
 
 import cn.stylefeng.guns.core.common.annotion.BussinessLog;
 import cn.stylefeng.guns.core.common.annotion.Permission;
 import cn.stylefeng.guns.core.common.constant.Const;
-import cn.stylefeng.guns.core.common.constant.dictmap.CompanyDict;
 import cn.stylefeng.guns.core.common.constant.dictmap.CustomerSiteDict;
-import cn.stylefeng.guns.core.common.constant.factory.ConstantFactory;
-import cn.stylefeng.guns.core.common.exception.BizExceptionEnum;
 import cn.stylefeng.guns.core.common.page.LayuiPageFactory;
-import cn.stylefeng.guns.core.log.LogObjectHolder;
 import cn.stylefeng.guns.core.shiro.ShiroKit;
-import cn.stylefeng.guns.modular.attendance.entity.AttendanceAllRecord;
-import cn.stylefeng.guns.modular.attendance.service.AttendanceService;
-import cn.stylefeng.guns.modular.company.wrapper.CompanyWrapper;
-import cn.stylefeng.guns.modular.system.entity.Company;
-import cn.stylefeng.guns.modular.system.entity.CustomerSite;
-import cn.stylefeng.guns.modular.system.model.CustomerSiteDto;
-import cn.stylefeng.guns.modular.system.model.UserDto;
-import cn.stylefeng.guns.modular.system.service.CustomerSiteService;
+import cn.stylefeng.guns.modular.company.entity.CustomerSite;
+import cn.stylefeng.guns.modular.company.service.CompanyService;
 import cn.stylefeng.guns.modular.system.warpper.CustomerSiteWrapper;
-import cn.stylefeng.guns.modular.system.warpper.UserWrapper;
 import cn.stylefeng.roses.core.base.controller.BaseController;
 import cn.stylefeng.roses.core.datascope.DataScope;
 import cn.stylefeng.roses.core.reqres.response.ResponseData;
-import cn.stylefeng.roses.kernel.model.exception.ServiceException;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import com.google.common.collect.Maps;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -66,7 +49,7 @@ public class CustomerSiteController extends BaseController {
     private String PREFIX = "/modular/customer/customer_site/";
 
    @Autowired
-   private CustomerSiteService customerSiteService;
+   private CompanyService.CustomerSiteService customerSiteService;
