+++ src/main/java/cn/stylefeng/guns/modular/employee/service/InfoMgrService.java
@@ -1,27 +1,15 @@
 package cn.stylefeng.guns.modular.employee.service;
 
-import java.util.List;
 import java.util.Map;
 
-import javax.annotation.Resource;
-
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 
-import cn.stylefeng.guns.core.common.exception.BizExceptionEnum;
 import cn.stylefeng.guns.core.common.page.LayuiPageFactory;
-import cn.stylefeng.guns.modular.system.entity.Dict;
 import cn.stylefeng.guns.modular.system.entity.InfoMgr;
-import cn.stylefeng.guns.modular.system.entity.Myself;
-import cn.stylefeng.guns.modular.system.mapper.CompanyMapper;
-import cn.stylefeng.guns.modular.system.mapper.DictMapper;
 import cn.stylefeng.guns.modular.system.mapper.InfoMgrMapper;
-import cn.stylefeng.guns.modular.system.mapper.MyselfMapper;
-import cn.stylefeng.roses.core.util.ToolUtil;
-import cn.stylefeng.roses.kernel.model.exception.ServiceException;
 
 /**
  * <p>
