--- src/main/java/cn/stylefeng/guns/modular/system/mapper/mapping/CompanyMapper.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="cn.stylefeng.guns.modular.system.mapper.CompanyMapper">
-
-	<!-- 通用查询映射结果 -->
-    <resultMap id="BaseResultMap" type="cn.stylefeng.guns.modular.system.entity.Company">
-        <id column="CUSTOMER_ID" property="customerID"/>
-        <result column="COMPANY_NAME" property="companyName"/>
-        <result column="COMPANY_ADDRESS" property="companyAddress"/>
-        <result column="TEL" property="tel"/>
-        <result column="MAIL" property="mail"/>
-        <result column="CEO_NAME" property="ceoName"/>
-        <result column="CEO_TEL" property="ceoTel"/>
-        <result column="CEO_MAIL" property="ceoMail"/>
-        <result column="CREATE_TIME" property="createTime"/>
-        <result column="CREATE_USER" property="createUser"/>
-        <result column="UPDATE_TIME" property="updateTime"/>
-        <result column="UPDATE_USER" property="updateUser"/>
-    </resultMap>
-    
-	<!-- 通用查询结果列 -->
-    <sql id="Base_Column_List">
-  			CUSTOMER_ID as customerID,
-        	COMPANY_NAME as companyName,
-			COMPANY_ADDRESS as companyAddress,
-			TEL as tel,
-			MAIL as mail,
-			CEO_NAME as ceoName,
-			CEO_TEL as ceoTel,
-			CEO_MAIL as ceoMail,
-			CREATE_TIME as createTime,
-			CREATE_USER as createUser,
-			UPDATE_TIME as updateTime,
-			UPDATE_USER as updateUser
-    </sql>
-    <select id="list" resultType="map">
-        select
-        <include refid="Base_Column_List"></include>
-        from customer
-        <if test="condition != null and condition != ''">
-            where COMPANY_NAME like CONCAT('%',#{condition},'%') 
-               or COMPANY_ADDRESS like CONCAT('%',#{condition},'%')
-        </if>
-        order by CREATE_TIME DESC
-    </select>
-
-</mapper>
