+++ src/main/webapp/pages/modular/customer/customer/company_add_site.html
@@ -1,22 +1,22 @@
 @layout("/common/_container.html",{bg:"bg-white",js:["/assets/modular/customer/customer/company_add_site.js"]}){
 <form id="companyForm" lay-filter="companyForm" class="layui-form model-form">
-    <!--input name="companyId" type="hidden"/-->
+    <input id="ipt-add-customer-id"name="customerID" type="hidden"/>
     <div class="layui-form-item">
         <label class="layui-form-label">现场名称<span style="color: red;">*</span></label>
         <div class="layui-input-block">
-            <input id="customersiteName" name="customersiteName" placeholder="请输入客户现场名称" type="text" class="layui-input" lay-verify="required" required/>
+            <input id="customerSiteName" name="customerSiteName" placeholder="请输入客户现场名称" type="text" class="layui-input" lay-verify="required" required/>
         </div>
     </div>
     <div class="layui-form-item">
         <label class="layui-form-label">现场地址<span style="color: red;">*</span></label>
         <div class="layui-input-block">
-            <input id="customersiteAddress" name="customersiteAddress" placeholder="请输入客户现场地址" type="text" class="layui-input" lay-verify="required" required autocomplete="off"/>
+            <input id="customerSiteAddress" name="customerSiteAddress" placeholder="请输入客户现场地址" type="text" class="layui-input" lay-verify="required" required autocomplete="off"/>
         </div>
     </div>
     <div class="layui-form-item">
         <label class="layui-form-label">现场车站<span style="color: red;">*</span></label>
         <div class="layui-input-block">
-            <input id="customersiteStation" name="customersiteStation" placeholder="请输入客户现场车站" type="text" class="layui-input" lay-verify="required" required/>
+            <input id="customerSiteStation" name="customerSiteStation" placeholder="请输入客户现场车站" type="text" class="layui-input" lay-verify="required" required/>
         </div>
     </div>
     <div class="layui-form-item text-right">
